% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiprobable_probability.R
\name{equiprobable_probability}
\alias{equiprobable_probability}
\title{Probability of Equiprobable Events}
\usage{
equiprobable_probability(positive_cases, total_cases)
}
\arguments{
\item{positive_cases}{Number of positive cases for the event.}

\item{total_cases}{Total number of cases.}
}
\value{
The probability of the event occurring.
}
\description{
This function calculates the probability of an event occurring given the number
of positive cases for that event and the total number of cases.
}
\examples{
equiprobable_probability(3, 10)
# Output: 0.3
}
