% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_algebra.R
\name{is_algebra}
\alias{is_algebra}
\title{Check if a collection of sets is an algebra}
\usage{
is_algebra(universe, sets)
}
\arguments{
\item{universe}{The sample space (vector)}

\item{sets}{A list of sets (list of vectors)}
}
\value{
TRUE if sets is an algebra, FALSE otherwise
}
\description{
Check if a collection of sets is an algebra
}
